+++ examples/typescript/simple_transfer.ts
@@ -4,20 +4,11 @@
  * This example shows how to use the Aptos client to create accounts, fund them, and transfer between them.
  */
 
-import {
-  Account,
-  AccountAddress,
-  Aptos,
-  AptosConfig,
-  InputViewFunctionJsonData,
-  Network,
-  NetworkToNetworkName,
-} from "@aptos-labs/ts-sdk";
+import { Account, AccountAddress, Aptos, AptosConfig, Network, NetworkToNetworkName } from "@aptos-labs/ts-sdk";
 import dotenv from "dotenv";
 
 dotenv.config();
 
-const APTOS_COIN = "0x1::aptos_coin::AptosCoin";
 const ALICE_INITIAL_BALANCE = 1_000_000_000;
 const BOB_INITIAL_BALANCE = 100;
 const TRANSFER_AMOUNT = 100;
@@ -25,24 +16,18 @@ const TRANSFER_AMOUNT = 100;
 // Default to devnet, but allow for overriding
 const APTOS_NETWORK: Network = NetworkToNetworkName[process.env.APTOS_NETWORK ?? Network.DEVNET];
 
-/**
- * Prints the balance of an account
- * @param aptos
- * @param name
- * @param address
- * @returns {Promise<*>}
- *
- */
-const balance = async (aptos: Aptos, name: string, address: AccountAddress): Promise<any> => {
-  const payload: InputViewFunctionJsonData = {
-    function: "0x1::coin::balance",
-    typeArguments: ["0x1::aptos_coin::AptosCoin"],
-    functionArguments: [address.toString()],
-  };
-  const [balance] = await aptos.viewJson<[number]>({ payload });
-
-  console.log(`${name}'s balance is: ${balance}`);
-  return Number(balance);
+const balance = async (
+  aptos: Aptos,
+  name: string,
+  address: AccountAddress,
+  versionToWaitFor?: bigint,
+): Promise<number> => {
+  const amount = await aptos.getAccountAPTAmount({
+    accountAddress: address,
+    minimumLedgerVersion: versionToWaitFor,
+  });
+  console.log(`${name}'s balance is: ${amount}`);
+  return amount;
 };
 
 const example = async () => {
@@ -80,8 +65,6 @@ const example = async () => {
   const aliceBalance = await balance(aptos, "Alice", alice.accountAddress);
   const bobBalance = await balance(aptos, "Bob", bob.accountAddress);
 
-  console.log("aliceBalance", aliceBalance);
-  console.log("ALICE_INITIAL_BALANCE", ALICE_INITIAL_BALANCE);
   if (aliceBalance !== ALICE_INITIAL_BALANCE) throw new Error("Alice's balance is incorrect");
   if (bobBalance !== BOB_INITIAL_BALANCE) throw new Error("Bob's balance is incorrect");
 
@@ -89,8 +72,7 @@ const example = async () => {
   const txn = await aptos.transaction.build.simple({
     sender: alice.accountAddress,
     data: {
-      function: "0x1::coin::transfer",
-      typeArguments: [APTOS_COIN],
+      function: "0x1::aptos_account::transfer",
       functionArguments: [bob.accountAddress, TRANSFER_AMOUNT],
     },
   });